Handle incomplete notebook toolbar contexts
Fix restart-kernel failures caused by VS Code serializing notebook
toolbar invocations with an empty notebook editor during lifecycle
transitions.

Accept the discriminated toolbar hint and fall back to the active
notebook only when its URI is omitted. Keep explicit targets strict, and
colocate notebook invocation decoding and resolution while attaching the
typed handler to the restart command.
