+++ extension/src/commands/publishMarimoNotebook.ts
@@ -1,6 +1,6 @@
 import { Effect } from "effect";
 
-import type { NotebookToolbarContext } from "../commands.ts";
+import type { NotebookToolbarContext } from "./NotebookCommandTarget.ts";
 import { publishMarimoNotebookGist } from "./publishMarimoNotebookGist.ts";
 
 export const publishMarimoNotebook = Effect.fn("command.publishMarimoNotebook")(
