+++ extension/src/commands/runStale.ts
@@ -1,11 +1,13 @@
 import { Effect, flow, Option } from "effect";
 
-import type { NotebookCommandTarget } from "../commands.ts";
 import { CellExecutions } from "../kernel/CellExecutions.ts";
-import { getNotebookCommandEditor } from "../lib/getNotebookCommandEditor.ts";
 import { showErrorAndPromptLogs } from "../lib/showErrorAndPromptLogs.ts";
 import { VsCode } from "../platform/VsCode.ts";
 import { MarimoNotebookDocument } from "../schemas/MarimoNotebookDocument.ts";
+import {
+  getNotebookCommandEditor,
+  type NotebookCommandTarget,
+} from "./NotebookCommandTarget.ts";
 
 export const runStale = Effect.fn("command.runStale")(
   function* (context?: NotebookCommandTarget) {
